backend/requirements.txt
fastapi==0.114.1
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic==2.8.2
SQLAlchemy==2.0.32
psycopg2-binary>=2.9.9
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
PyJWT==2.9.0

pymupdf               # PyMuPDF
pdfplumber            # optional alt extractor
python-docx
pillow
pytesseract

boto3                 # S3
google-cloud-storage  # GCS

email-validator
python-multipart
